C++11: Modernized the RandomPicker
Replaced rand() with std::default_random_engine and used
std::uniform_real_distribution to generate random float or double
numbers. Also made the real type a template parameter, to avoid some
implicit float / double conversion warnings.

+++ src/plugins/gmx/gmxplugin.cpp
@@ -393,7 +393,7 @@ bool GmxPlugin::write(const Map *map, const QString &fileName)
                         bgName = tileset->name();
 
                         int xInTilesetGrid = tile->id() % tileset->columnCount();
-                        int yInTilesetGrid = (int)(tile->id() / tileset->columnCount());
+                        int yInTilesetGrid = tile->id() / tileset->columnCount();
 
                         xo = tileset->margin() + (tileset->tileSpacing() + tileset->tileWidth()) * xInTilesetGrid;
                         yo = tileset->margin() + (tileset->tileSpacing() + tileset->tileHeight()) * yInTilesetGrid;
